+++ tests/sdk/v2/test_viz.py
@@ -14,6 +14,7 @@ class TestWfDefGraph:
     @staticmethod
     def test_example_nodes_edges():
         """ """
+        aggregate_step_name = "aggregation_step"
         wf_def = sample_wfs.wf(123).model
         graph = _viz.wf_def_graph(wf_def)
 
@@ -38,6 +39,7 @@ class TestWfDefGraph:
                 caption=["tests.sdk.v2.sample_wfs.add():12", "invocation-3-task-add"],
             ),
         ]
+
         assert graph.nodes2 == [
             _viz.Node(id="artifact-0-inc", caption=[]),
             _viz.Node(id="artifact-1-inc-2", caption=[]),
@@ -47,6 +49,7 @@ class TestWfDefGraph:
             _viz.Node(id="constant-0", caption=["3"]),
             _viz.Node(id="constant-1", caption=["123"]),
             _viz.Node(id="constant-2", caption=["6"]),
+            _viz.Node(id=aggregate_step_name, caption=["outputs"]),
         ]
         assert graph.edges == [
             ("artifact-1-inc-2", "invocation-0-task-inc"),
@@ -60,4 +63,6 @@ class TestWfDefGraph:
             ("artifact-4-integer-division", "invocation-3-task-add"),
             ("constant-2", "invocation-3-task-add"),
             ("invocation-3-task-add", "artifact-3-add"),
+            ("artifact-3-add", aggregate_step_name),
+            ("artifact-0-inc", aggregate_step_name),
         ]