+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 🐛 *Bug Fixes*
 * Retry getting results from CE if the results were not ready but the workflow succeeded.
 * Using secrets inside the workflow function will now work correctly on Ray
+* Fix WorkflowDef.graph - honour kwargs of tasks and add aggregate_output to show outputs
 
 💅 *Improvements*